+++ lib/cogs/exp.py
@@ -6,43 +6,41 @@ from discord import Member, Embed
 from discord.ext.commands import Cog
 from discord.ext.commands import CheckFailure
 from discord.ext.commands import command, has_permissions
-
 from discord.ext.menus import MenuPages, ListPageSource
-
+from lib.bot import bot
 from ..db import db # pylint: disable=relative-beyond-top-level
 
-"""
-FIX THE XP LEADERBOARD BELOW
-"""
+# daniel was here :o
 
-# class Menu(ListPageSource):
-#     def __init__(self, ctx, data):
-#         self.ctx = ctx
+class Menu(ListPageSource):
+    def __init__(self, ctx, data):
+        self.ctx = ctx
+        self.bot = bot
 
-#         super().__init__(data, per_page=10)
+        super().__init__(data, per_page=10)
 
-#     async def write_page(self, menu, offset, fields=[]):
-#         offset = (menu.current_page*self.per_page) + 1
-#         len_data = len(self.entries)
+    async def write_page(self, menu, offset, fields=[]):
+        offset = (menu.current_page*self.per_page) + 1
+        len_data = len(self.entries)
 
-#         embed = Embed(title="XP Leaderboard", description="See who is on top!", colour=self.ctx.author.colour)
-#         embed.set_thumbnail(url=self.ctx.guild.me.avatar_url)
-#         embed.set_footer(text=f"{offset:,} - {min(len_data, offset+self.per_page-1):,} of {len_data:,} members.")
+        embed = Embed(title="XP Leaderboard", description="See who is on top!", colour=self.ctx.author.colour)
+        embed.set_thumbnail(url=self.ctx.guild.me.avatar_url)
+        embed.set_footer(text=f"{offset:,} - {min(len_data, offset+self.per_page-1):,} of {len_data:,} members.")
         
-#         for name, value in fields:
-#             embed.add_field(name=name, value=value, inline=False)
+        for name, value in fields:
+            embed.add_field(name=name, value=value, inline=False)
 
-#         return embed
+        return embed
 
-#     async def format_page(self, menu, entries):
-#         offset = (menu.current_page*self.per_page) + 1
-#         fields = []
-#         table = ("\n".join(f"{idx+offset}. {self.ctx.guild.get_member(entry[0]).display_name} (XP: {entry[1]} | Level: {entry[2]})"
-#                    for idx, entry in enumerate(entries))) # FIX SOON
+    async def format_page(self, menu, entries):
+        offset = (menu.current_page*self.per_page) + 1
+        fields = []
+        table = ("\n".join(f"{idx+offset}. {bot.get_user(entry[0]).name} (XP: {entry[1]} | Level: {entry[2]})"
+                   for idx, entry in enumerate(entries)))
 
-#         fields.append(("Ranks", table))
+        fields.append(("Ranks", table))
 
-#         return await self.write_page(menu, offset, fields)
+        return await self.write_page(menu, offset, fields)
 
 class Exp(Cog):
     def __init__(self, bot):
@@ -92,27 +90,24 @@ class Exp(Cog):
             db.commit()
             await ctx.send(f"Level messages set to {yes_or_no}.")
 
-    """
-    FIX THE XP LEADERBOARD BELOW
-    """
-    # @command(name="leaderboard", aliases=["lb", "xplb"])
-    # async def display_leaderboard(self, ctx):
-    #     records = db.records("SELECT UserID, XP, Level FROM exp ORDER BY XP DESC")
+    @command(name="leaderboard", aliases=["lb", "xplb"])
+    async def display_leaderboard(self, ctx):
+        records = db.records("SELECT UserID, XP, Level FROM exp ORDER BY XP DESC")
 
-    #     menu = MenuPages(source=Menu(ctx, records), clear_reactions_after=True, timeout=100.0)
-    #     await menu.start(ctx)
+        menu = MenuPages(source=Menu(ctx, records), clear_reactions_after=True, timeout=100.0)
+        await menu.start(ctx)
 
-    @command(name="leaderboard", aliases=["lb", "xplb"], brief="The XP Leaderboard is under construction.")
-    async def display_leaderboard_error(self, ctx):
-        prefix_thing = db.field("SELECT Prefix FROM guilds WHERE GuildID = ?", ctx.guild.id)
-        embed = Embed(title="Doob XP Leaderboard is currently under construction.", 
-                     description=f"You can view your global xp rank by doing {str(prefix_thing)}level, sorry for the inconvinence.",
-                     colour=ctx.author.colour,
-                     timestamp=datetime.utcnow())
+    # @command(name="leaderboard", aliases=["lb", "xplb"], brief="The XP Leaderboard is under construction.")
+    # async def display_leaderboard_error(self, ctx):
+    #     prefix_thing = db.field("SELECT Prefix FROM guilds WHERE GuildID = ?", ctx.guild.id)
+    #     embed = Embed(title="Doob XP Leaderboard is currently under construction.", 
+    #                  description=f"You can view your global xp rank by doing {str(prefix_thing)}level, sorry for the inconvinence.",
+    #                  colour=ctx.author.colour,
+    #                  timestamp=datetime.utcnow())
         
-        embed.set_thumbnail(url=ctx.guild.me.avatar_url)
+    #     embed.set_thumbnail(url=ctx.guild.me.avatar_url)
 
-        await ctx.send(embed=embed)
+    #     await ctx.send(embed=embed)
 
     @Cog.listener()
     async def on_ready(self):